Use mixins to extend the functionality of the base class
calculatorMixin and randomizerMixin are functions that take a base class (Foo) and return new classes.
The classes returned by these mixins extend the functionality of the provided base class by adding the calc() method in calculatorMixin and the randomize() method in randomizerMixin.
Bar is a new class that extends Foo and combines both the functionalities of calculatorMixin and randomizerMixin.